+++ src/main/java/ojosama/talkak/category/domain/Category.java
@@ -1,6 +1,8 @@
-package ojosama.talkak.keyword.model;
+package ojosama.talkak.category.domain;
 
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -14,12 +16,13 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Getter
-@Table(name = "keyword")
-public class Keyword {
+@Table(name = "category")
+public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private String keyword;
+    @Enumerated(EnumType.STRING)
+    private CategoryType categoryType;
 
 }
