+++ src/main/java/ojosama/talkak/category/domain/CategoryType.java
@@ -1,9 +1,6 @@
-package ojosama.talkak.category.model;
+package ojosama.talkak.category.domain;
 
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
 import java.util.Arrays;
-import java.util.stream.Stream;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
