+++ src/main/java/ojosama/talkak/category/domain/MemberCategory.java
@@ -1,11 +1,8 @@
-package ojosama.talkak.category.model;
+package ojosama.talkak.category.domain;
 
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
@@ -16,7 +13,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import ojosama.talkak.common.exception.TalKakException;
 import ojosama.talkak.common.exception.code.MemberError;
-import ojosama.talkak.member.model.Member;
+import ojosama.talkak.member.domain.Member;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
