+++ src/main/java/ojosama/talkak/category/repository/CategoryRepository.java
@@ -1,10 +1,8 @@
 package ojosama.talkak.category.repository;
 import java.util.Optional;
-import ojosama.talkak.category.model.Category;
-import ojosama.talkak.category.model.CategoryType;
+import ojosama.talkak.category.domain.Category;
+import ojosama.talkak.category.domain.CategoryType;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
