+++ src/main/java/ojosama/talkak/keyword/domain/Keyword.java
@@ -1,8 +1,6 @@
-package ojosama.talkak.category.model;
+package ojosama.talkak.keyword.domain;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -16,13 +14,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Getter
-@Table(name = "category")
-public class Category {
+@Table(name = "keyword")
+public class Keyword {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Enumerated(EnumType.STRING)
-    private CategoryType categoryType;
+    private String keyword;
 
 }
