+++ src/main/java/ojosama/talkak/reaction/repository/ReactionRepository.java
@@ -1,7 +1,7 @@
 package ojosama.talkak.reaction.repository;
 
-import ojosama.talkak.reaction.model.Reaction;
-import ojosama.talkak.reaction.model.ReactionId;
+import ojosama.talkak.reaction.domain.Reaction;
+import ojosama.talkak.reaction.domain.ReactionId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
