+++ src/main/java/ojosama/talkak/video/controller/VideoController.java
@@ -4,12 +4,12 @@ import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.List;
-import ojosama.talkak.video.dto.VideoRequest;
-import ojosama.talkak.video.dto.VideoResponse;
-import ojosama.talkak.video.dto.YoutubeApiResponse;
-import ojosama.talkak.video.dto.YoutubeCategoryRequest;
-import ojosama.talkak.video.dto.YoutubeUrlValidationRequest;
-import ojosama.talkak.video.dto.YoutubeUrlValidationResponse;
+import ojosama.talkak.video.request.VideoRequest;
+import ojosama.talkak.video.response.VideoResponse;
+import ojosama.talkak.video.response.YoutubeApiResponse;
+import ojosama.talkak.video.request.YoutubeCategoryRequest;
+import ojosama.talkak.video.request.YoutubeUrlValidationRequest;
+import ojosama.talkak.video.response.YoutubeUrlValidationResponse;
 import ojosama.talkak.video.service.AwsS3Service;
 import ojosama.talkak.video.service.VideoService;
 import ojosama.talkak.video.service.YoutubeService;
