+++ src/main/java/ojosama/talkak/video/repository/VideoRepository.java
@@ -1,6 +1,6 @@
 package ojosama.talkak.video.repository;
 
-import ojosama.talkak.video.model.Video;
+import ojosama.talkak.video.domain.Video;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
