+++ src/main/java/ojosama/talkak/video/request/AwsS3Request.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.video.dto;
+package ojosama.talkak.video.request;
 
 public record AwsS3Request(
         String filename
