+++ src/main/java/ojosama/talkak/video/request/VideoRequest.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.video.dto;
+package ojosama.talkak.video.request;
 
 public record VideoRequest(String title, Long memberId, Long categoryId) {
