+++ src/main/java/ojosama/talkak/video/request/YoutubeCategoryRequest.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.video.dto;
+package ojosama.talkak.video.request;
 
 public record YoutubeCategoryRequest(Long categoryId) {
