+++ src/main/java/ojosama/talkak/video/response/AwsS3Response.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.video.dto;
+package ojosama.talkak.video.response;
 
 public record AwsS3Response(
         String url
