+++ src/main/java/ojosama/talkak/video/response/VideoResponse.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.video.dto;
+package ojosama.talkak.video.response;
 
 public record VideoResponse(Long id, String title, Long memberId, Long categoryId, String uniqueFileName) {
