+++ src/main/java/ojosama/talkak/video/response/YoutubeApiResponse.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.video.dto;
+package ojosama.talkak.video.response;
 import java.time.LocalDateTime;
 
 public record YoutubeApiResponse(LocalDateTime date, String videoId, String channelId, String title, String thumbnailUrl) {
