+++ src/main/java/ojosama/talkak/video/service/AwsS3Service.java
@@ -7,16 +7,14 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import ojosama.talkak.category.repository.CategoryRepository;
 import ojosama.talkak.common.exception.TalKakException;
 import ojosama.talkak.common.exception.code.VideoError;
-import ojosama.talkak.video.dto.AwsS3Request;
-import ojosama.talkak.video.dto.AwsS3Response;
-import ojosama.talkak.video.dto.VideoRequest;
-import ojosama.talkak.video.dto.VideoResponse;
-import ojosama.talkak.video.model.Video;
+import ojosama.talkak.video.request.AwsS3Request;
+import ojosama.talkak.video.response.AwsS3Response;
+import ojosama.talkak.video.request.VideoRequest;
+import ojosama.talkak.video.response.VideoResponse;
+import ojosama.talkak.video.domain.Video;
 import ojosama.talkak.video.repository.VideoRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
