+++ src/main/java/ojosama/talkak/video/service/YoutubeService.java
@@ -8,8 +8,8 @@ import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 import ojosama.talkak.category.repository.CategoryRepository;
-import ojosama.talkak.video.dto.YoutubeApiResponse;
-import ojosama.talkak.video.dto.YoutubeCategoryRequest;
+import ojosama.talkak.video.response.YoutubeApiResponse;
+import ojosama.talkak.video.request.YoutubeCategoryRequest;
 import ojosama.talkak.common.util.WebClientUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
