+++ src/test/java/ojosama/talkak/member/service/MyPageServiceTest.java
@@ -6,10 +6,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import ojosama.talkak.category.model.Category;
-import ojosama.talkak.category.model.CategoryType;
+import ojosama.talkak.category.domain.Category;
+import ojosama.talkak.category.domain.CategoryType;
 
-import ojosama.talkak.category.model.MemberCategory;
+import ojosama.talkak.category.domain.MemberCategory;
 import ojosama.talkak.category.repository.CategoryRepository;
 import ojosama.talkak.category.repository.MemberCategoryRepository;
 import ojosama.talkak.common.exception.TalKakException;
@@ -17,9 +17,9 @@ import ojosama.talkak.common.exception.code.MemberError;
 import ojosama.talkak.member.dto.MyPageInfoRequest;
 import ojosama.talkak.member.dto.MyPageInfoResponse;
 import ojosama.talkak.member.dto.MyPageInfoResponse.CategoryResponse;
-import ojosama.talkak.member.model.Age;
-import ojosama.talkak.member.model.Member;
-import ojosama.talkak.member.model.MembershipTier;
+import ojosama.talkak.member.domain.Age;
+import ojosama.talkak.member.domain.Member;
+import ojosama.talkak.member.domain.MembershipTier;
 import ojosama.talkak.member.repository.MemberRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
